DC-13: Plugin registry, BibTeX cross-source, assessment fixes — version 1.0.1
Plugin registry pattern replaces 6 repetitive init blocks with data-driven
factory map. BibTeX journal extraction now checks all source-specific metadata
keys. Fixed errors.Is() comparison, dead format converter, HuggingFace E2E
race condition, version test helper synchronization, sort modernization.

+++ internal/rtv.plugin.arxiv.go
@@ -387,7 +387,7 @@ func (p *ArXivPlugin) doRequest(ctx context.Context, reqURL string) (*arxivFeed,
 	}
 
 	var feed arxivFeed
-	limitedBody := io.LimitReader(resp.Body, arxivMaxResponseBytes)
+	limitedBody := io.LimitReader(resp.Body, int64(arxivMaxResponseBytes))
 	if err := xml.NewDecoder(limitedBody).Decode(&feed); err != nil {
 		return nil, fmt.Errorf("%w: %w", ErrArxivXMLParse, err)
 	}
